pyproject.toml
[project]
name = "lil-rag"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.4"
dependencies = [
    "langchain-community==0.2.16",
    "langchain==0.2.16",
    "langchainhub==0.1.15",
    "tiktoken==0.7.0",
    "bs4==0.0.2",
    "lxml==5.3.0",
    "watchdog==4.0.1",
    "pytest==8.2.1",
    "python-dotenv==1.0.1",
    "pypdf==4.2.0",
    "pypdf2==3.0.0",
    "pytest-asyncio==0.24.0",
    "email-validator==2.2.0",
    "pydantic[email]>=2.10.4",
    "python-multipart==0.0.9",
    "argon2-cffi==23.1.0",
    "passlib==1.7.4",
    "pyjwt==2.9.0",
    "instructor==1.4.0",
    "sqlglot==25.25.1",
    "cohere==5.9.0",
    "boto3==1.34.131",
    "amazon_textract_caller==0.2.3",
    "amazon-textract-textractor==1.7.12",
    "Pillow==10.3.0",
    "pandas==2.2.2",
    "numpy==1.26.4",
    "umap-learn==0.5.6",
    "fake-useragent==1.5.1",
    "alembic==1.13.2",
    "psycopg2-binary==2.9.10",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio,asyncpg]>=2.0.37",
    "slowapi==0.1.9",
    "requests==2.32.3",
    "uvicorn==0.30.6",
    "fastapi==0.112.1",
    "pydantic-ai[logfire]>=0.0.19",
    "opentelemetry-instrumentation-asgi>=0.50b0",
    "logfire[fastapi]>=3.2.0",
    "typing-extensions>=4.12.2",
    "pytest-mock>=3.14.0",
    "transformers>=4.48.2",
    "openai>=1.59.8",
    "pgvector>=0.3.6",
]
